debug: fixed product routes
