+++ server/api/index.js
@@ -8,17 +8,8 @@ const isLoggedIn = (req, res, next) => {
   next()
 }
 
-const isAdmin = (req, res, next) => {
-  if (!req.user || !req.user.isAdmin) {
-    const err = new Error("You're not an admin!")
-    err.status = 401
-    return next(err)
-  }
-  next()
-}
-
-router.use('/users', isAdmin, require('./users'))
-router.use('/products', isAdmin, require('./products'))
+router.use('/users', require('./users'))
+router.use('/products', require('./products'))
 router.use('/orders', isLoggedIn, require('./orders'))
 
 router.use((req, res, next) => {
