+++ server/api/products.js
@@ -25,7 +25,16 @@ router.get('/:productId', async (req, res, next) => {
   }
 })
 
-router.post('/', async (req, res, next) => {
+const isAdmin = (req, res, next) => {
+  if (!req.user || !req.user.isAdmin) {
+    const err = new Error("You're not an admin!")
+    err.status = 401
+    return next(err)
+  }
+  next()
+}
+
+router.post('/', isAdmin, async (req, res, next) => {
   try {
     const newProduct = await Product.create(req.body)
     res.json(newProduct)
@@ -34,7 +43,7 @@ router.post('/', async (req, res, next) => {
   }
 })
 
-router.put('/:productId', async (req, res, next) => {
+router.put('/:productId', isAdmin, async (req, res, next) => {
   try {
     const currProduct = await Product.findByPk(req.params.productId)
     if (currProduct) {
@@ -48,7 +57,7 @@ router.put('/:productId', async (req, res, next) => {
   }
 })
 
-router.delete('/:productId', async (req, res, next) => {
+router.delete('/:productId', isAdmin, async (req, res, next) => {
   try {
     const currProduct = await Product.findByPk(req.params.productId)
     if (currProduct) {
